labels being hidden and unhidden
